Invent adb_hba_file_rules_view view to show the content of pg_hba.conf.
This view is designed to wit it shows what is currently in the file, not what
the postmaster has loaded as the active settings.  That allows it to be used to
pre-vet edits before issuing SIGHUP.  As with the earlier view, go out of our
way to allow errors in the file to be reflected in the view, to assist that
use-case.

(We might at some point invent a view to show the current active settings,
but this is not that patch; and it's not trivial to do.)

Haribabu Kommi, reviewed by Ashutosh Bapat, Michael Paquier, Simon Riggs,
and myself

Discussion: https://postgr.es/m/CAJrrPGerH4jiwpcXT1-46QXUDmNp2QDrG9+-Tek_xC8APHShYw@mail.gmail.com

This is partial backport de16ab7 by Georgy Shelkovy <g.shelkovy@arenadata.io>
without backporting catalog changes and handling changes to SSL settings in
version 10