Fix explain analyze error handling
Error handling for explain analyze differs from a regular query.
When we execute a plan slice and get an error in a case on a simple
query we just cleanup and rethrow an exception out of PG_TRY block.
But for explain analyze we tried to return stats to QD though it is
possibly incorrect or even broken. It can cause different errors
with stats packets during:
* serialization on QE ("InstrEndLoop called on running node")
* deserialization on QD ("Invalid execution statistics received
stats node-count mismatch")

Also in a case of a double fault an original error was overridden
by a new one. It hid an original error from a user and this
behavior differed from a simple query plan processing.