Fix multiple definition linker error in unit tests
In the same spirit as ee7eb0e8ec77d1b6, this removes the extra
definition in postgres_test.c.

This is uncovered by building with GCC 10, where -fno-common is the new
default [1][2] (vis a vis -fcommon). I could also reproduce this by
turning on "-fno-common" in older releases of GCC and Clang.

Backpatch to 6X_STABLE.