Fix preprocessing of NOT IN to LAS-Apply in filter
NOT IN is equivalent to "<> ALL" SQL expression and should be
replaced with LAS-Apply in CSubqueryHandler::FRemoveAllSubquery().
Before current commit ORCA transformed "<> ALL" subquery in filter
incorrectly. A query

select * from R where R.a <> ALL (select S.b from S)

was translated to

select * from R LAS-APPLY-Not-In (select S.b from S where S.b = R.a) on true;

This plan returns wrong results when S.b is NULL. To handle it
(S.b = R.a) should be checked with IS_NOT_FALSE expression.

For example in https://github.com/greenplum-db/gpdb/issues/10967

create table t1(a int) distributed by (a);
insert into t1 values (1), (2), (3);
select * from t1 where a not in (
    select b.a from t1 a left join t1 b on false
);

returned incorrect (1), (2), (3) result on ORCA and correct empty
result on postgres planner. The root of it was in FRemoveAllSubquery().
Previously this function transformed

+--CScalarSubqueryAll(<>)["a" (16)]   origin: [Grp:6, GrpExpr:0]
    |--CLogicalLeftOuterJoin   origin: [Grp:4, GrpExpr:0]
    |  |--CLogicalGet "t1" ("t1"), Columns: ["a" (8),..]
    |  |--CLogicalConstTableGet Columns: ["a" (16),..]
    |  +--CScalarConst (1)   origin: [Grp:3, GrpExpr:0]
    +--CScalarIdent "a" (0)   origin: [Grp:5, GrpExpr:0]

to

+--CLogicalLeftAntiSemiApplyNotIn (Reqd Inner Cols: "a" (16))
    |--CLogicalGet "t1" ("t1"), Columns: ["a" (0),..]
    |--CLogicalSelect
    |  |--CLogicalLeftOuterJoin   origin: [Grp:4, GrpExpr:0]
    |  |  |--CLogicalGet "t1" ("t1"), Columns: ["a" (8),..]
    |  |  |--CLogicalConstTableGet Columns: ["a" (16),..]
    |  |  +--CScalarConst (1)   origin: [Grp:3, GrpExpr:0]
    |  +--CScalarCmp (=)
    |     |--CScalarIdent "a" (0)   origin: [Grp:5, GrpExpr:0]
    |     +--CScalarIdent "a" (16)
    +--CScalarConst (1)

and didn't add IS_NOT_FALSE check as we should expect.
Now we get a correct logical tree with empty result:

+--CLogicalLeftAntiSemiApplyNotIn (Reqd Inner Cols: "a" (16))
    |--CLogicalGet "t1" ("t1"), Columns: ["a" (0),..]
    |--CLogicalSelect
    |  |--CLogicalLeftOuterJoin   origin: [Grp:4, GrpExpr:0]
    |  |  |--CLogicalGet "t1" ("t1"), Columns: ["a" (8),..]
    |  |  |--CLogicalConstTableGet Columns: ["a" (16),..]
    |  |  +--CScalarConst (1)   origin: [Grp:3, GrpExpr:0]
    |  +--CScalarIsDistinctFrom (=)
    |     |--CScalarCmp (=)
    |     |  |--CScalarIdent "a" (0)   origin: [Grp:5, GrpExpr:0]
    |     |  +--CScalarIdent "a" (16)
    |     +--CScalarConst (0)
    +--CScalarConst (1)