Fix partition selector under join by keys with different types.
This fix regards only projection-based implementation of partition
selector that works only under postgres planner. The plan has to have
the following fragment:
    Join
     Join Cond: partitioned_table.id = othertable.id
      -> Append
         -> Result
            One-Time Filter: PartSelected
               -> Seq Scan partition1
         -> Result
            One-Time Filter: PartSelected
               -> Seq Scan partition2
         ...
      -> Some materialization point
         -> PartitionSelector for partitioned_table
            Filter: othertable.id
            -> SeqScan othertable
               Filter: <other conditions>

Executor under postgres planner performs projection of input tuple to
further determine partitions appropriate for dynamic scan node. At the
same time, such partition selection mechanics assumes that the type of
search tuple after projection has to match to the schema of partitioning
table and, as a result, search values inside this tuple must have the
same types as in partitioning key. This is so despite of internal
comparator used inside partition definition routines allows to compare
values with compatible but different types.
Perhaps, this assumption originates from the fact that expressions in
the target list of projection have to coerce variables of incoming tuple
into values of the same type as at attributes of partitioning key. But
there are some data types, e.g. timestamp and date, that are compared
directly (there is equality function for them) and aren't casted to the
same representation. For these data types planner doesn't build casting
functions over attributes and comparator function have to be picked up
based on these types. Therefore, the assumption about types of result
values after projection equaled to the types of partitioning key is not
correct and have to be fixed.

The current patch removes restriction of projection-based partition
selector execution node so that it would be able to apply search values
having types different from partitioning key. Also patch contains some
refactoring for the second strategy of dynamic partition selection
routine based on execution of partitioning level expressions (ORCA
case). In particular, it specifies for tupdesc argument the role of
tuple descriptor for transferable values and isnull arguments as it's
done in the first projection-based strategy. Also it removes exprTypeOid
argument as unnecessary.

Backport to 5X from commit 999d3f27c5.