Demonstrate CVE-2020-25695 on Greenplum
The idea of CVE-2020-25695 is to make superuser fire a defered
trigger with security invoker function that executes some
malicious code with superuser privileges.

We create a two-stap trap for a superuser running a maintenance
operation (ANALYZE) on a "trap" table. First of all we
create an index with stable function and replace it with volatile
one after creation. This index function is invoked by ANALYZE and
acts as a vulnerability firestarter - it inserts data to some table
aside ("executor"). "Executor" table has a defered insert trigger
with security invoker function that executes malicious code under
superuser. In a current example we give superuser to our unprivileged
user on segments (not on coordinator node).

Current test should fail and demonstrate that unprivileged user
becomes superuser on segments.