Don't try to generate generic plans with GPORCA.
GPORCA doesn't support Param nodes, so it will always fall back to the Postgres planner. What happened without this patch was that the backend code would compare the cost of the custom plan generated with GPORCA with the cost of a generic plan generated with the Postgres planner, and that doesn't make much sense because the GPORCA has a very different cost model from the Postgres planner.

This commit is a rework of 09aa23d32c433dfbf9675286799027f5867484b9 with additional 'is_custom_cost_safe' flag. It indicates is it safe to compare costs of custom and generic plan (true value), was custom plan ever been generated using ORCA (false value).
Unlike mentioned commit, this one contains two additional tests, which proves
 - ORCA's custom plan is always chosen instead of Postgres generic plan
 - Switch to generic plan is possible even if ORCA is enabled (fallback case)