Refactor stack unwind in GPDB
ORCA and GPDB relied on frame registers to unwind the stack. This
approach required -fno-omit-frame-pointer during compilation.
Though it works faster than a backtrace(), the downside is reserving
a register by the compiler that affects the result code's performance.
Also, it is not an easy thing to port ORCA to other platforms (we
need to write assembly-specific code for that).

Besides, gp_backtrace() has some problems:

1. fallback to backtrace() from frame pointers was not SIGSEGV-safe
   for the builds without -fno-omit-frame-pointer.
2. backtrace() is known to be async-unsafe on the first run, while
   all the next runs are async-safe. For example on Linux backtrace()
   first run calls dlopen() to load libgcc that evaluates malloc().
   So we need to make an initial call of backtrace() before using it
   in a signal handler.

So, this commit removes the need to use -fno-omit-frame-pointer and
make all the code use backtrace() in a safe manner.

Co-authored-by: Ivan Leskin <leskin.in@phystech.edu>