Convert float8 with double instead of long double
It seems that GPDB faced a glibc bug on RedHat 6 with denormalized
numbers and fixed it using two-step conversion in float8in():

- text to long double
- long double to double after denormalized numbers validation

This differs from PG approach where we rely on libc and make direct
conversion to double. GPDB two-step way faced a problem on PowerPC
architecture: it doesn't have analogue of Intel 80-bit long double
with 63-bit mantissa and extended range between 1.0E-4932W and
1.0E+4932W. Instead PowerPC has and extended double type combined
from two doubles with 106-bit mantissa and the same range as double
provides: 1.0E−323L - 1.0E+308L. As a result we can't use
the same validation logic for the numbers near +/- infinity as it
is possible to do on Intel. So this GPDB specific code doesn't help
us to solve the problem on RedHat 6 but gives us a performance penalty:
we have extended double hardware support only since PowerPC 7 with
VSX instructions and can benefit from it on new versions of GCC.
So, lets get rid of it and return the code to PG approach with double.