Backport synchronization of amgetbitmap definition with upstream
GPDB has a different definition of amgetbitmap with upstream.
Since GPDB also supports StreamBitmap in bitmap index. So normally
we need to create a specific bitmap node in the amgetbitmap AM.
We used to return the created node, this leads to losing the tuples
matched in the bitmap.
So I refactor the amgetbitmap function definition to set bitmap
through pointer and then return the tuples matched in the bitmap.

(backported from 3245d17937bf990ea00689770ec068b12e400eb8)