Fixup snapshoot_eof output
A copy-past inaccuracy was found in snapshot_eof test. It uses
preprocessing of the isolation framework to unify tests for row
and column oriented tables. So in two places "column" orientation
was used instead of "@orientation@" template. It didn't affect
test itself as it is about manipulations with fault injector under
"ignore". Though it is the only test uses this fault and this
inaccuracy doesn't influence any other test, it can become a problem
in future. So, fix it.