Clarify IS DISTINCT FROM FALSE three valued logic (#1239)
Commit e4173271a1f5e250b87ed37e02c8a462627d4029 fixed three valued
logic for IS DISTINCT FROM FALSE transformation of the LASHJ to HJ.

The problem was, that this commit created too general restriction
for the inner and outer sides of the LASJ: they had to be exactly
NOT NULL for any type of LASJ. But it is redundant for LASJ Not-In
case, as it returns nothing if the inner-outer tuple match status
is NULL.

The reason we have to care about this clarification is that we have
our own implementation of a preprocessing fix for NOT NULL in LASJ
filter (90afebcd9404a87ab1c4e95b17a6f6a2734b4d63). It differs form
the Pivotal's realization (ee9dc5d21c7e2f4b38dfa80900f56ff518047acf)
and adds IS DISTINCT FROM FALSE for joins during transformation.
Before current patch some joins produced non-effective plans due
to e4173271a1f5e250b87ed37e02c8a462627d4029.

(cherry picked from commit bbd42d35cf5b88085c2f9de0834513c514d35689)