Enable back log_lock_waits GUC
Commit ffd8acc have disabled log_lock_waits GUC because of
non-consistence of ResProcSleep() with ProcSleep() routine in case of
temporary semaphore wakeup logic to trigger logging of lock waits. In
particular, in ResProcSleep() there is omitted a loop around semaphore
lock that forces to resuspend main thread after logging of current lock
info if deadlock is not detected. Current commit adds this loop and
other neccessary stuff to make possible to handle wakeups that prints
out lock info.