Add -fsigned-char flag for compilaction on power
This commit introduces
- a new compiler option -fsigned-char aimed to align the default char
behavior on PowerPC and x86 (default char is unsigned on Power and is
signed on x86).
- fixes in the functions that work with 'firstchar' byte value received
from socket. This is done for consistency across the codebase and
avoidance of further misreads of -1 (EOF) value in case gp is compiled
with default char unsigned (it can mistakenly be treated as 255 then).