added add_nested_subplans
By the time we get here, the plan already has dedicated subplans for each partition. The intention of this part of adjust_appendrel_attrs_mutator is to create a copy of the most-closely-nested subplan and add it to subplans list. It's implied that nested subplans of a copy will be added later in fixup_subplans, but it's the cause of a crush: more nested subplans will be added after their parent subplansin a flat list. The intension of this fix is to recursively traverse through all the subplans and add them right on the spot. This will make the order of subplans correct and will leave nothing to do for fixup_subplans.