fix lock upgrade issue for part tables with index
planner chooses lock level for relation indexes depending on kind
of operation on table: Row Exclusive for DML and Access Share
for SELECT. The kind of operation is detected by
root->parse->resultRelation that contains only a single rel index.
inheritance planner mutate query tree for each partition before
building a plan for it. We should handle this issue in the same
manner to avoid lock upgrading that may lead to deadlocks.