Exclude gpmmon process from shmem users
If we terminate the session from gpmmon using the pg_terminate_backend
function, gpprefmon background process exits with failure exit code.
As a result, the postmaster instance goes into recovery mode. It was
found out that the problem was the process was marker as shared memory user.
I made sure that the process is not using shared memory and disabled
the BGWORKER_SHMEM_ACCESS flag solved the gpmon crash problem.