Log skips of files during resync
The pg_aocsseg table was not resynced during incremental recovery. There is the
"resync TID" record in log, but there is no corresponding "resync scheduled"
record. Logging of all possible skip situations is added to understand the
cause of the error. Now this logging is unconditional.
The file name logging error is fixed. Before the patch the previous file name or
uninitialized buffer was passed to the elog function.