Fix high memory consumption for queries with SQL subqueries
Steps to reproduce:
1. Create function written in sql language that optimizers can’t inline, e. g.:
create function f(a int) returns int
immutable
security definer
language sql as $$
  select case when $1 > 500 then 0 end;
$$;
2. Execute a query which executes this function multiple times
explain analyze verbose
select f(i) from generate_series(1, 1000000) i;
3. The query uses much more memory when gp_enable_gpperfmon=true than in the
opposite case.

The reason of high memory consumption is memory allocation for two structures in
CreateQueryDesc() and release of memory for only one of these structures in
FreeQueryDesc(). This pair of functions is run every time when subquery is
executed. The more times the subquery is executed, the more memory leaks.

The patch adds deallocation of the second structure after data in it is no
longer needed. There is no leak in _SPI_execute_plan(), but deallocation is
added here to be sure that the leak would not appear in the future.