Add --exclude="db_analyze" to tar of template (#476)
If run `analyze_db` for any databse, than for this database folder
"db_analyze" is created. It may contain large volume of data.
If after it create new segment using `gpexpand`, than folder "db_analyze"
will be copied to new segment. This folder is useless for new segment.
Because of this option --exclude="db_analyze" is added to command of tar.

Also new behave test on gpexpand was added to check that gpexpand does not
copy extra directories.