Move pg_stat_statements query jumbling to core
Moving of jumbling logic to core part gives the opportunity to reuse this logic
in extensions without code duplication. This patch is a partial backport of
https://github.com/postgres/postgres/commit/5fd9dfa5f50e4906c35133a414ebec5b6d518493

The differences between this patch and the original one are:
- add release of JumbleState memory;
- there is no compute_query_id GUC;
- queryId computation algorithm is not changed, so the CleanQuerytext() function
and the querytext argument of JumbleQuery() are not created;
- the JumbleQuery() function is called from pg_stat_statements.c, so no changes
to the hook arguments are required.

The original patch author is Julien Rouhaud, its reviewers are Alvaro Herrera,
Nitin Jadhav and Zhihong Yu.