Fix segment segfault when join inner plan contains outer params.
When a subquery has WHERE clause with equality condition between the
relation's distribution key and params, which refer to the relations from
upper query levels (for example, lateral reference in join subquery), then
the distiribution key's EquivalenceClass may contain these params. Therefore,
these params may be put into the subplan node's targetList. There are the
cases when such behaviour leads to these params are being referenced in
different slices, and it probably could lead to segfault when the executor
is accessing its attribute list. A typical case is shown below:

```
create table t1( t1c1 text ) distributed by (t1c1);
create table t2( t2c1 text ) distributed by (t2c1);
explain verbose select * from t2 join lateral
(select * from t1 where t2.t2c1 = t1.t1c1 limit 1) x on true;
                            QUERY PLAN
-----------------------------------------------------------------------
 Nested Loop
   Output: t2.t2c1, t1.t1c1
   ->  Gather Motion 3:1  (slice1; segments: 3)
          Output: t2.t2c1
         ->  Seq Scan on public.t2
               Output: t2.t2c1
   ->  Materialize
          Output: t1.t1c1
         ->  Limit
                Output: t1.t1c1
               ->  Result
                      Output: t1.t1c1
                     Filter: ((t2.t2c1) = t1.t1c1)
                     ->  Materialize
                            Output: t1.t1c1, (t2.t2c1)
                           ->  Gather Motion 3:1  (slice2; segments: 3)
                                  Output: t1.t1c1, (t2.t2c1)
                                 ->  Seq Scan on public.t1
                                        Output: t1.t1c1, t2.t2c1

```
The above plan is invalid because Seq Scan on t1 is performed at slice2
and uses params that are scanned from t2 at slice1. And the presence of
such params may lead to segfault during inner plan execution.

This commit fixes the bug by changing the cdbpullup_findEclassInTargetList
function without breaking its inherent behaviour and purpose. The call of the
function takes place when the search through the given equivalence class
for a member, which is already inside the targetList or uses no Vars that are
not in targetList, is required. In query above it is used to fill the Flow's
hashExprs list during plan creation for the bottom Seq Scan. The root of the
problem in case above is that hashExprs contains entry of type T_Param,
which refer to outer relations, and this param is appended to final targetList
from the hashExprs. In this case the hashExprs list is filled with the equivalence
class member of type T_Param, which appears with the WHERE clause. The proposed
solution makes the cdbpullup_findEclassInTargetList function iterate through whole
equivalence class and through whole targetList first, hoping to find member, which
is equal to targetList entry. And only next step is to check whether
member is pseudoconst or non-Var that is not in targetList, because for
given query Param is added due to these conditions. These changes prevent subplan
from holding outer refs in similar cases.