Don't initialize ORCA on query executors
Currently, the ORCA optimizer is initialized unconditionally in all backends.
But we can reduce memory consumption by avoiding ORCA initialization
on query executors where ORCA is never used.
This can be done using the Gp_role variable, which is set to GP_ROLE_EXECUTE
on the executors, and this happens before ORCA is initialized.