Add support for EXPLAIN ANALYZE queries containing DECLARE CURSOR
An error occurred while executing a query like "EXPLAIN ANALYZE DECLARE CURSOR"
because segments tried to open a cursor. This behavior was due to the fact
that in the query plan on segments the utilityStmt field was not NULL
and contained the DeclareCursorStmt node.

After this commit the utilityStmt field is set to NULL in the dispatched query
plan to properly process the query plan on segments.

The ExplainOnePlan() function comment says that the utilityStmt field should be
ignored and the query from the cursor should be execute. Therefore, fixes
have been added in this function.

When executing DECLARE CURSOR without EXPLAIN, in function PerformCursorOpen()
the utilityStmt field is discarded after the plan is built.
This is also mentioned in the comment to the PerformCursorOpen() function.
This fix is similar - we discard the utilityStmt field after the query plan
is built. We do not copy the plan before set utilityStmt field to NULL since
the utilityStmt field is not needed anywhere else.