Fix ORCA SubLink processing at query normalization
The ORCA optimizer may produce an invalid query plan or fallback to the legacy
optimizer due to incorrect processing of nested SubLink entries during query
normalization.
Firstly, the issue occurs when a query has a multi-level SubLink entry in its
targetList, and this SubLink contains attributes that are referenced in the
query's GROUP BY clause. Currently, ORCA may construct for such queries an
invalid plan, according to which some wrong and unexpected attributes get into
the nested part of SubLink's plan (example of such query is provided in
tests section). And the presence of those attributed may lead to incorrect query
execution or even segmentation fault.

The ORCA during its work constructs a DXL representation of the Query tree, and
in order to make the final tree as simple as possible it normalizes original
Query structures via CQueryMutators::NormalizeQuery function. Inside the
NormalizeQuery function the NormalizeGroupByProjList function is called, where
the problems, that are considered in this patch, occur. In details, inside
NormalizeGroupByProjList the ConvertToDerivedTable is called, which returns a
new query that wraps the original query as rtable entry. Then the original
(derived) query's targetList is being iterated, and for entries, which are
grouping columns, FixGroupingCols function is called, where each targetEntry
is being mutated with RunGroupingColMutator mutator function. Then this entry
is added to a mutator context's (SContextGrpbyPlMutator) targetList
(context.m_lower_table_tlist) and also replaces the Expr for the entry in
original targetList to make it reference the newly added entry in the
m_lower_table_list. Therefore, because not all entries of original targetList
are grouping columns and not all are being mutated, the
context.m_lower_table_tlist may become shorter than original. After iteration
process, the entries of original targetList, that are not grouping columns,
are modified by RunExtractAggregatesMutator function, which checks whether the
entry contains attributes from grouping clause and replaces them with
corresponding expression from context.m_lower_table_tlist. Here, during
replacement, the original expression's varattno is assigned to
context.m_lower_table_tlist entry resno, also the varlevelsup is also being
changed corresponding to the current query level. And if these values will be
incorrect or remain unmodified, the ORCA may built invalid plan or fallback to
legacy optimizer. This happens because the new context.m_lower_table_tlist is
then being saved in derived query and, as it was said before, it can have
length and order of elements different from original, and, therefore, if the
mutation process is wrong, the entries of modified original targetList may not
correspond to the right entries of new context.m_lower_table_tlist, which leads
to invalid plan construction. (there is a mechanism of hash maps, if they are
filled wrongly due to wrong targetLists, the mapping vars -> relation column
will be incorrect).

When a query contains nested SubLinks, RunExtractAggregatesMutator mutates
only the upper level of nested structure, and other subqueries remain
unmodified, because RunExtractAggregatesMutator does not have a branch for the
Query, which is required for futher diving into lower subqueries.

This patch modifies RunExtractAggregatesMutator function by adding the Query
branch to make the mutator correctly step into nested SubLink queries. And the
call of MutateQueryOrExpressionTree is replaced to the call of mutator itself
in order to correctly modify m_current_query_level when stepping into lower
subqueries.

Secondly, the ORCA may fallback to the legacy optimizer or build invalid plan,
when nested SubLink entry is in GROUP BY clause itself (see the test
section). In this case nested SubLinks are mutated incorrectly.
As it was mentioned earlier, inside the NormalizeGroupByProjList function the
derived query's targetList is being iterated, and for entries, which are
grouping columns, FixGroupingCols function is called, where each targetEntry
is being mutated with RunGroupingColMutator mutator function.
RunGroupingColMutator handles multi-level SubLink incorrectly as well. After
stepping into MutateQueryOrExpressionTree inside the SubLink branch and
mutating upper query level, the mutation on lower query rte is called,
and RunGroupingColMutator is called for this level. It has Query node type to
process, however, when diving into second level deep SubLink subquery, the
m_current_query_level variable is not increased, because
MutateQueryOrExpressionTree, which is responsible for calling mutator on the
lower subquery, does not know anything about a context. Therefore, for the
query like in the tests, that behaviour leads to extra modification of the Var's
varlevelsup value. And when Var's varlevelsup value is modified, it starts
referencing different query level. This may lead to the fallback or invalid plan
construction when accessing attributes from hash maps.

This patch fixes erroneous behaviour by replacing the call of
MutateQueryOrExpressionTree to the call of RunGroupingColMutator in order to
properly modify m_current_query_level when mutating lower level rtes.

Thirdly, the case of correlated SubLink under the AggRef is considered.
Such queries has simple structure like SELECT (SELECT max((SELECT ... FROM
s WHERE t.a=s.b))) ... FROM t. And for this case the ORCA may fallback
as well because of erroneous normalization algorithm.
In this case the NormalizeGroupByProjList is called anyway, because the query
has aggregates. While mutating the SubLink inside RunExtractAggregatesMutator,
the Aggref branch is executed for SubLink's query targetList, which mutates
args of the AggRef. And if a Var has varlevelsup value greater than the
agglevelsup, i.e the Var references a relation that is higher than the AggRef,
the Var is not mutated at all. The varlevelsup field is not modified and because
of that, when AggRef is pulled up to the derived query level (by appending it
into the context.m_lower_table_tlist inside MakeVarInDerivedTable), the Var,
whose varlevelsup was higher than AggRef's, will reference the wrong relation.
This may lead to the fallback or invalid plan construction when accessing
attributes from hash maps.

This patch makes the mutator function modify the varlevelsup of the Var
relatively the AggRef level in order to preserve proper query level.