Rewrite test interrupt_holdoff_count (#574)
This commit fixes test, which added at commit a5d36791329a80ad3471808e166c0bb05ee7c0d0

At first version of this test there was `declare cursor`.
QD does not wait result of `declare cursor`, but on segments there is work,
which ends with error. This error is processed when `abort` is called.
At some cases `abort` is called before error is generated on segment.
It led to situation, when there is not any errors on segments and `abort`
is executed with success result. This test requiers error to check
`InterruptHoldoffCount` after query which ended with error.

New version of the test uses existing inject fault, which generate error at `create table`.
`create table` ends with error and `InterruptHoldoffCount` can be checked correctly.

Co-authored-by: Vasiliy Ivanov <ivi@arenadata.io>