Fix incorrect plan when using CTE with gp_cte_sharing enabled.
The problem is that when creating a T_ModifyTable node, the targetlist is not specified for it, and therefore the targetlist is empty in the parent nodes T_Material and T_ShareInputScan, which leads to an error.

Fixed by specifying targetlist after creating T_ModifyTable node in subquery_planner function. targetlist is retrieved from the returningList of the Query structure. Also removing the Assert that the targetlist is empty and filling targetlist it in the set_plan_references function since it is now specified at an earlier stage.