ORCA initialization refactoring
When initializing ORCA mixed a C approach (returning error codes and then
checking them) and a C++ approach (throwing exceptions and then catching them).
This mixing led to errors. For example, in case of low memory or any other
exception, the CWorkerPoolManager::Init function did not return GPOS_OK, and
then the gpos_init function called
the CMemoryPoolManager::GetMemoryPoolMgr()->Shutdown function, which reset
CMemoryPoolManager::m_memory_pool_mgr. In this case, all the exceptions that
arose were not rethrown, so the execution continued, and the following
gpdxl_init function was called, in which the CAutoMemoryPool class was
initialized. In the constructor of this class, the static method
CMemoryPoolManager::GetMemoryPoolMgr returned NULL, because we nullified
m_memory_pool_mgr up the stack. Thus, a segfault could occur.

The patch replaces C approach with C++ approach in ORCA initialization and adds
processing of standard C++ exceptions, which are thrown for example by the new
operator in case of out of memory.

Now a backend terminates with error on any exception during ORCA initialization,
because ORCA does not work correctly when initialization is not fully completed.

CAutoMemoryPool has been replaced with this class constructor content in the
places, where CAutoMemoryPool destructor could be called after an exception. In
this cases we could get exception in CAutoMemoryPool destructor during
processing of other exception. This caused to backend termination without
informative message.

The gpdxl_init function has been changed, because the auto memory pool is used
to automatically free memory when it goes out of scope, but here the pools were
immediately removed, and in fact, they are not needed here at all.